Significant changes to the way widgets work.
Setting them up so they import templates at macro time based on either
metadata (inline or file-import), or a file with a similar name to the class.
(eg. app.MyWidget becomes app/MyWidget.html)

Currently seems to be loading the templates correctly, which for me and my macro
creating experience, is a big deal :)