Update for haxe3 - get
In Haxe2 (and neko1.8) the macro parser for Xml ignored whitespace,
which was really convenient - it cleaned up our JS output significantly
and resulted in a smaller file size.

In Haxe3, the parser is much more accurate, and keeps all the whitespace.
I've modified our widget processor to emulate the old behaviour - because
I think it was preferable, and because it means our unit tests will still pass :)

Unit tests currently passing on JS, Neko, CPP.

Flash still failing because of weird issues moving elements that are already
part of the DOM...