Changed behaviour of boolean dtx attributes.
These were previously undocumented and unspecified, and i changed my
mind on their behaviour.

Now usage is like this:

    <div dtx-show="true">Always show</div>
    <div dtx-show="false">Never show</div>
    <div dtx-show="someBoolean">Show if `someBoolean` is true</div>
    <div dtx-hide="someBoolean">Hide if `someBoolean` is true</div>
    <div dtx-show="someString!=null">$someString</div> // Only show if someString is not null
    <div dtx-show="someArray.length>0">...</div> // Only show if someArray is not empty

The difference though is that now you must explicitly declare your boolean attributes.  This will
break some of my code, but as it was undocumented, untested and unspecified, I'm happy to change it.

Doing this also resulted in me fixing issues with inline variable initialization, (public var x=0;)
which is working correctly and unit tested now.