Major refactor of the way named elements and partials work
Before, every time you an element needs to be known at runtime, it would
add a data-dtx-id attribute, and then do a find() at runtime.

Apart from being slow, this created issues when multiple widgets with the same
named partials occupied the same DOM space - it would match the element in both
widgets, not just one.

The solution is to map all of the referenced elements in the widget constructor,
save them in an array reference, and then simply refer to the array reference.

This is faster and more reliable.

I've added new unit tests to verify the behaviour, and they're all passing on neko.