Set partial variables from call-site in parent template
Done in a pretty nice way, any variables from the parent template are found,
and the updates are added to the setter.  So every time the parent variable is set,
the partial variable is updated accordingly.  Can work with complex expressions too.

Also cleaned up some code, more reliable parsing of partials before the rest of the code,
and not parsing the children of a partial declaration.  Before we were getting some partial
variables also declared in the parent template etc, not any more.

As a result of these changes, partials must now be declared on the top level, which is good practice anyway.