More null handling in string interpolation.
Previously we covered it if it was a class field, now we also cover it
if it's a new variable we create a field for (and there is no print function),
or if there is a method call.  Method calls with null values are not called,
they are replaced with "".